An example to get you started.
Here is an example of some RSA exchange code. I added some notes of things that should be changed. It would also be cool if you had two different parties actually enter numbers and allowed a third party (attempted attacker) to watch the public exchange and attempt to guess the private keys. (This could be done with a simple input() or actually do some server work and let all parties use a different device in different locations. I recommend FLASK in Python, PHP, or NODE for that.) Such as what happens in a real life RSA examples like SSL/TLS.